Se mueven las funciones a panel_utils.py para dejar limpio el fdtd
